refactored server file, added order blockage for unlogged in students
